+++ betfairmarketvolumesclient/utils.py
@@ -9,10 +9,15 @@ from betfairmarketvolumesclient.resources.marketresource import MarketSelection
 logger = logging.getLogger(__name__)
 
 
-def parse_from_data_frame(data: DataFrame) -> List[MarketSelection]:
+def parse_from_data_frame(
+    data: DataFrame, file_url: str, country: str
+) -> List[MarketSelection]:
     if isinstance(data, DataFrame):
         try:
-            return [MarketSelection(**d) for d in data.to_dict("records")]
+            return [
+                MarketSelection(**{**d, **{"_file_url": file_url, "country": country}})
+                for d in data.to_dict("records")
+            ]
         except Exception as e:
             logger.error(f"Issue parsing selections, {e}")
     return []
@@ -31,6 +36,6 @@ def build_greyhound_url(base_url: str, win_or_place: str, date: datetime) -> str
 
 
 def build_horse_racing_url(
-        base_url: str, country: str, win_or_place: str, date: datetime
+    base_url: str, country: str, win_or_place: str, date: datetime
 ) -> str:
     return f"{base_url}prices{country}{win_or_place}{date.strftime('%d%m%Y')}.csv"